Moved files to database module
