+++ light-crud-api/src/database/mod.rs
@@ -1,0 +1,4 @@
+pub mod frame;
+pub mod frame_data;
+pub mod initialize;
+pub mod location;
